change Proof inner representation from Vec<u32> to Vec<u8>, align proof type with recent p2p spec

+++ crates/p2p_proto/src/transaction.rs
@@ -87,7 +87,7 @@ pub struct InvokeV3 {
 #[protobuf(name = "crate::proto::transaction::InvokeV3WithProof")]
 pub struct InvokeV3WithProof {
     pub invoke: InvokeV3,
-    pub proof: Vec<u32>,
+    pub proof: Vec<u8>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, ToProtobuf, TryFromProtobuf, Dummy)]
